+++ mod-source-record-storage-server/src/test/java/org/folio/services/RecordServiceTest.java
@@ -310,6 +310,36 @@ public class RecordServiceTest extends AbstractLBServiceTest {
     });
   }
 
+  @Test
+  public void shouldGetSourceRecordsByListOfIds(TestContext context) {
+    Async async = context.async();
+    List<Record> records = TestMocks.getRecords();
+    RecordCollection recordCollection = new RecordCollection()
+      .withRecords(records)
+      .withTotalRecords(records.size());
+    recordService.saveRecords(recordCollection, TENANT_ID).onComplete(batch -> {
+      if (batch.failed()) {
+        context.fail(batch.cause());
+      }
+      List<String> ids = records.stream()
+        .map(record -> record.getExternalIdsHolder().getInstanceId())
+        .collect(Collectors.toList());
+      recordService.getSourceRecords(ids, ExternalIdType.INSTANCE.name(), TENANT_ID).onComplete(get -> {
+        if (get.failed()) {
+          context.fail(get.cause());
+        }
+        List<SourceRecord> expected = records.stream()
+          .map(RecordDaoUtil::toSourceRecord)
+          .collect(Collectors.toList());
+        Collections.sort(expected, (r1, r2) -> r1.getRecordId().compareTo(r2.getRecordId()));
+        Collections.sort(get.result().getSourceRecords(), (r1, r2) -> r1.getRecordId().compareTo(r2.getRecordId()));
+        context.assertEquals(expected.size(), get.result().getTotalRecords());
+        compareSourceRecords(context, expected, get.result().getSourceRecords());
+        async.complete();
+      });
+    });
+  }
+
   // TODO: test get source records between two dates
 
   @Test